+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 * **Fixed**
   * MCP230xx cleanup pin configured log entries.
+  * OneWire cleanup logging on device configured and during bus scan.
 
 ## 0.8.0